requirements.txt
djangorestframework==3.16.1
drf-yasg==1.21.11